Highlight instructions with {cond} at the very end
When trying to use e.g. ADR{cond}L in the unified syntax, the GNU
assembler warns that the conditional infixes are deprecated. The
preferred form is ADRL{cond} instead. It is also in line with arm.com
and keil.com documentation.

This applies to other instructions that have optional parts (e.g. "S"
for updating condition flags) - the condition goes last.